fixed sound warnings, tweaks to functionality
	modified:   ../libs/openFrameworks/sound/ofSoundStream.cpp
	modified:   ../libs/openFrameworks/sound/ofSoundStream.h
	modified:   ../libs/openFrameworks/sound/ofSoundUnit.cpp
	modified:   ../libs/openFrameworks/sound/ofSoundUnit.h
	modified:   ../libs/openFrameworks/sound/ofSoundEffect.h

+++ libs/openFrameworks/sound/ofSoundStream.cpp
@@ -69,7 +69,7 @@ vector<ofSoundSource*> ofSoundStreamGetAllSoundSources() {
 }
 
 
-long unsigned long ofSoundStreamGetCurrentTick() {
+unsigned long ofSoundStreamGetCurrentTick() {
 	return tickCount;
 }
